requirements.txt
boto3==1.34.162
psycopg2-binary==2.9.9
requests==2.32.4
dnspython==2.7.0 